What: fix name format issue with !go1.6
Why:

  * variable func's name formatted in /path/to/gogo.glob.func?

How:

  * nop
